Fix some issues and move newsletter to home
