+++ src/components/Dropdown/index.js
@@ -1,0 +1,3 @@
+import Dropdown from './Dropdown'
+
+export default Dropdown
