+++ src/components/LanguageSelection/index.js
@@ -1,0 +1,3 @@
+import LanguageSelection from './LanguageSelection'
+
+export default LanguageSelection
